Prepare for python3 (this is python2 code that 2to3 can automatically convert)
Note: the save-file and preferences-file handling isn't fixed yet, so
it doesn't support saving, and the resolution preference resets on
each restart.

After using the 2to3 tool to convert this to a python3-pygame, some
hand editing was needed; mainly it's checking which of the integer
divisions need to be converted back to integers. This commit is the
result of fixing by hand after running 2to3, and then cherry-picking
those changes back to the python2 version.