Prepare file handling for Python 3
This commit is python2 code, but the project can now be converted by
the 2to3 tool to be python3 code, with no further hand-editing
required.

Note: python3 can load the saves generated by python2, but python2
can't load a python3 pickle, so all the config will be lost (and the
savefiles be unreadable) if you run lightyears with python2 after
running it with python3.